refactor: move opengl service types to a new namespace
