+++ Testing/VelaptorTests/Content/TextureTests.cs
@@ -15,6 +15,7 @@ using Velaptor.Content;
 using Velaptor.Factories;
 using Velaptor.Graphics;
 using Velaptor.NativeInterop.OpenGL;
+using Velaptor.NativeInterop.Services;
 using Velaptor.OpenGL;
 using Velaptor.ReactableData;
 using Xunit;
