+++ Testing/VelaptorTests/OpenGL/Buffers/FontGpuBufferTests.cs
@@ -17,6 +17,7 @@ using Velaptor;
 using Velaptor.Factories;
 using Velaptor.Graphics;
 using Velaptor.NativeInterop.OpenGL;
+using Velaptor.NativeInterop.Services;
 using Velaptor.OpenGL;
 using Velaptor.OpenGL.Batching;
 using Velaptor.OpenGL.Buffers;
