+++ Testing/VelaptorTests/OpenGL/Buffers/GpuBufferBaseTests.cs
@@ -17,6 +17,7 @@ using Moq;
 using Velaptor;
 using Velaptor.Factories;
 using Velaptor.NativeInterop.OpenGL;
+using Velaptor.NativeInterop.Services;
 using Velaptor.OpenGL;
 using Velaptor.OpenGL.Batching;
 using Velaptor.OpenGL.Buffers;
