+++ Testing/VelaptorTests/OpenGL/Buffers/LineGpuBufferTests.cs
@@ -20,6 +20,7 @@ using NSubstitute;
 using Velaptor;
 using Velaptor.Factories;
 using Velaptor.NativeInterop.OpenGL;
+using Velaptor.NativeInterop.Services;
 using Velaptor.OpenGL;
 using Velaptor.OpenGL.Batching;
 using Velaptor.OpenGL.Buffers;
