+++ Testing/VelaptorTests/OpenGL/Buffers/ShapeGpuBufferTests.cs
@@ -22,6 +22,7 @@ using Velaptor;
 using Velaptor.Factories;
 using Velaptor.Graphics;
 using Velaptor.NativeInterop.OpenGL;
+using Velaptor.NativeInterop.Services;
 using Velaptor.OpenGL;
 using Velaptor.OpenGL.Batching;
 using Velaptor.OpenGL.Buffers;
