+++ Testing/VelaptorTests/OpenGL/Shaders/ShaderProgramTests.cs
@@ -14,6 +14,7 @@ using Moq;
 using Velaptor;
 using Velaptor.Factories;
 using Velaptor.NativeInterop.OpenGL;
+using Velaptor.NativeInterop.Services;
 using Velaptor.OpenGL;
 using Velaptor.OpenGL.Exceptions;
 using Velaptor.OpenGL.Services;
