+++ Testing/VelaptorTests/OpenGL/Shaders/ShapeShaderTests.cs
@@ -15,6 +15,7 @@ using Moq;
 using Velaptor;
 using Velaptor.Factories;
 using Velaptor.NativeInterop.OpenGL;
+using Velaptor.NativeInterop.Services;
 using Velaptor.OpenGL;
 using Velaptor.OpenGL.Services;
 using Velaptor.OpenGL.Shaders;
