+++ Testing/VelaptorTests/OpenGL/Shaders/TextureShaderTests.cs
@@ -16,6 +16,7 @@ using Moq;
 using Velaptor;
 using Velaptor.Factories;
 using Velaptor.NativeInterop.OpenGL;
+using Velaptor.NativeInterop.Services;
 using Velaptor.OpenGL;
 using Velaptor.OpenGL.Services;
 using Velaptor.OpenGL.Shaders;
