+++ Velaptor/OpenGL/Buffers/FontGpuBuffer.cs
@@ -15,6 +15,7 @@ using Exceptions;
 using Factories;
 using GpuData;
 using NativeInterop.OpenGL;
+using NativeInterop.Services;
 using ReactableData;
 
 /// <summary>
