+++ Velaptor/OpenGL/Buffers/GpuBufferBase.cs
@@ -9,6 +9,7 @@ using System.Diagnostics.CodeAnalysis;
 using Carbonate.Fluent;
 using Factories;
 using NativeInterop.OpenGL;
+using NativeInterop.Services;
 using ReactableData;
 
 /// <summary>
