+++ Velaptor/OpenGL/Buffers/LineGpuBuffer.cs
@@ -14,6 +14,7 @@ using ExtensionMethods;
 using Factories;
 using GpuData;
 using NativeInterop.OpenGL;
+using NativeInterop.Services;
 using ReactableData;
 
 /// <summary>
