+++ Velaptor/OpenGL/Buffers/ShapeGpuBuffer.cs
@@ -18,6 +18,7 @@ using Factories;
 using GpuData;
 using Graphics;
 using NativeInterop.OpenGL;
+using NativeInterop.Services;
 using ReactableData;
 
 /// <summary>
