+++ Velaptor/OpenGL/Shaders/FontShader.cs
@@ -8,6 +8,7 @@ using System;
 using Carbonate.Fluent;
 using Factories;
 using NativeInterop.OpenGL;
+using NativeInterop.Services;
 using ReactableData;
 using Services;
