+++ Velaptor/OpenGL/Shaders/ShaderProgram.cs
@@ -10,6 +10,7 @@ using Carbonate.Fluent;
 using Exceptions;
 using Factories;
 using NativeInterop.OpenGL;
+using NativeInterop.Services;
 using Services;
 
 /// <inheritdoc/>
